% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-fao.R
\name{.get_faostat_token}
\alias{.get_faostat_token}
\title{Fetch a guest token for the FAOSTAT API}
\usage{
.get_faostat_token(verbose = FALSE)
}
\value{
Character. A valid Bearer token string.
}
\description{
Fetch a guest token for the FAOSTAT API
}
\keyword{internal}
