% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-miscellaneous.R
\name{encode_hapi_app_identifier}
\alias{encode_hapi_app_identifier}
\title{Encode an HDX HAPI app identifier}
\usage{
encode_hapi_app_identifier(application, email)
}
\arguments{
\item{application}{Character. Calling application name.}

\item{email}{Character. Contact email address.}
}
\value{
Character. Encoded app identifier for HDX HAPI calls.
}
\description{
Encode an HDX HAPI app identifier
}
\examples{
\dontrun{
encode_hapi_app_identifier("my-analysis", "me@example.org")
}
}
