% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-miscellaneous.R
\name{get_hdx_hapi_availability}
\alias{get_hdx_hapi_availability}
\title{Get HDX HAPI data availability}
\usage{
get_hdx_hapi_availability(
  iso3 = NULL,
  category = NULL,
  subcategory = NULL,
  admin_level = NULL,
  ...
)
}
\arguments{
\item{iso3}{Character. Optional ISO3 country code.}

\item{category}{Character. Optional HAPI category.}

\item{subcategory}{Character. Optional HAPI subcategory.}

\item{admin_level}{Integer. Optional admin level.}

\item{...}{Additional arguments passed to \code{get_hdx_hapi()}.}
}
\value{
A tibble.
}
\description{
Get HDX HAPI data availability
}
