% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-miscellaneous.R
\name{get_hdx_hapi_food_security}
\alias{get_hdx_hapi_food_security}
\title{Get HDX HAPI food security data}
\usage{
get_hdx_hapi_food_security(
  iso3 = NULL,
  ipc_phase = NULL,
  ipc_type = NULL,
  start_date = NULL,
  end_date = NULL,
  admin_level = NULL,
  mrv = NULL,
  ...
)
}
\arguments{
\item{iso3}{Character. Optional ISO3 country code.}

\item{ipc_phase}{Character. Optional IPC phase.}

\item{ipc_type}{Character. Optional IPC type.}

\item{start_date, end_date}{Character or Date. Optional reference period
overlap filters.}

\item{admin_level}{Integer. Optional admin level.}

\item{mrv}{Integer. Optional number of most recent years to keep after
retrieval, based on \code{reference_period_start}.}

\item{...}{Additional arguments passed to \code{get_hdx_hapi()}.}
}
\value{
A tibble.
}
\description{
Get HDX HAPI food security data
}
