% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-miscellaneous.R
\name{get_hdx_hapi_population}
\alias{get_hdx_hapi_population}
\title{Get HDX HAPI baseline population data}
\usage{
get_hdx_hapi_population(
  iso3 = NULL,
  gender = NULL,
  age_range = NULL,
  admin_level = NULL,
  start_date = NULL,
  end_date = NULL,
  mrv = NULL,
  ...
)
}
\arguments{
\item{iso3}{Character. Optional ISO3 country code.}

\item{gender}{Character. Optional gender filter.}

\item{age_range}{Character. Optional age range filter.}

\item{admin_level}{Integer. Optional admin level.}

\item{start_date, end_date}{Character or Date. Optional reference period
overlap filters.}

\item{mrv}{Integer. Optional number of most recent years to keep after
retrieval, based on \code{reference_period_start}.}

\item{...}{Additional arguments passed to \code{get_hdx_hapi()}.}
}
\value{
A tibble.
}
\description{
Get HDX HAPI baseline population data
}
