% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-miscellaneous.R
\name{get_hdx_hapi_wfp_commodities}
\alias{get_hdx_hapi_wfp_commodities}
\title{Get HDX HAPI WFP commodities}
\usage{
get_hdx_hapi_wfp_commodities(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{get_hdx_hapi()}.}
}
\value{
A tibble.
}
\description{
Get HDX HAPI WFP commodities
}
