% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-miscellaneous.R
\name{get_hdx_hapi_wfp_markets}
\alias{get_hdx_hapi_wfp_markets}
\title{Get HDX HAPI WFP markets}
\usage{
get_hdx_hapi_wfp_markets(iso3 = NULL, as_sf = FALSE, ...)
}
\arguments{
\item{iso3}{Character. Optional ISO3 country code.}

\item{as_sf}{Logical. If TRUE, return an sf object using lon/lat coordinates.}

\item{...}{Additional arguments passed to \code{get_hdx_hapi()}.}
}
\value{
A tibble, or an sf object when \code{as_sf = TRUE}.
}
\description{
Get HDX HAPI WFP markets
}
