% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-miscellaneous.R
\name{list_osm_feature_classes}
\alias{list_osm_feature_classes}
\title{List OpenStreetMap Feature Classes}
\usage{
list_osm_feature_classes()
}
\value{
A character vector of feature class names.
}
\description{
Lists the feature classes available through \code{get_osm_feature_class()}.
}
\examples{
list_osm_feature_classes()
}
